feat(messages): image lightbox in the dashboard chat (shared with full page)
Tapping an image in the dashboard Messages card opened a new browser tab; the
full-page view had a proper full-screen lightbox (prev/next + Esc/←/→ +
download). Extract that Lightbox into a shared components/messages/Lightbox.tsx
and wire the dashboard's ThreadQuickView to it via the shared renderer's
onOpenImage hook, so both surfaces zoom images identically. SignalThreadView
now imports the shared Lightbox instead of its local copy.

tsc + eslint clean; 4 Lightbox tests + 23 messaging tests pass.
